Send columns individually and ensure proper types.
If we send all the columns together, the types seen in R are not
correct. In addition to sending them individually, we need to cast
numeric columns to be doubles as integer columns can be
problematic for plotting (possibly for other uses as well). Add
some test cases for sending data frames into R.